Changed Disk#size to parse the bytes value directly from the fdisk(8) output.
Parsing the converted value was causing an issue as fdisk was using SI units
and the rails methods were expecting binary units.
For example, 5MB was being converted to 5 * 1024 * 1024 bytes by rails
when fdisk intended the value to be 5 * 1000 * 1000 bytes.
This change avoids the confusion by parsing the bytes value directly.

https://bugzilla.redhat.com/show_bug.cgi?id=1096331