Allow NetworkInterface creation without a config file
This will allow a user on a Red Hat system to use the
"getter" methods defined in network_interface.rb even
if they don't have a config file corresponding to the
given interface name.
To do this we instantiate a NetworkInterfaceGeneric
object if the file doesn't exist rather than a
NetworkInterfaceRH instance.

Fixes #139