Add a #show method to the service classes
For systemd show prints some useful metrics about the unit and
the cgroup at runtime.

This can be used to get memory used, memory/cpu limits and the main
PID of the unit.
https://www.freedesktop.org/software/systemd/man/systemctl.html#show%20PATTERN...%7CJOB...

There is no equivalent call for SysV scripts so we raise
NotImplementedError for the show method there.