Add platform attribute to Scap class
Newer versions of scap-security-guide (> 0.1.32) add new xccdf
files which match our glob pattern, but do not contain the
remediations for the rules we want to run.

If we edit one of these files rather than the one for our
target platform, the rules will not be remediated properly.

Specifying the platform allows us to find the file we need.

https://bugzilla.redhat.com/show_bug.cgi?id=1493193