Prioritize SubscriptionManager over Rhn
Now the `LinuxAdmin::Rhn` has been deprecated:

    https://github.com/ManageIQ/linux_admin/pull/198
    1ifd7c4f788ce34200ff09ef56176397306904f85b

This means that every instantiation of `LinuxAdmin::Rhn` will then add a
deprecation warning to STDOUT.

Previously, this would also be triggered even if you used the public
class methods on `LinuxAdmin::SubscriptionManager` (the deprecation's
suggested replacement), because under the hood it determine which to
(`Rhn` or `SubscriptionManager`) by initializing each and calling their
`registered?` methods.

Furthermore, these means if `Rhn` is registered on the system in
addition to the preferred `SubscriptionManager`, it the
`LinuxAdmin::SubscriptionManager.validate_credentials` class method
would actually end up calling `LinuxAdmin::Rhn.validate_credentials`
instead because of the preference.

* * *

While this favoring should probably be refine to first look at the class
the `method_missing` method is being called against, this fix simply
favors `SubscriptionManager` over `Rhn` when deciding which class to
use.