Support nvme device partition naming
This commit collects the disk model from the parted output
and uses that to change how we construct the partition path.

Right now we keep the previous behavior of <device_path><partition_id>
in all cases except nvme when we use <device_path>p<partition_id>.
Previously, we would list a path which didn't exist for nvme devices.

https://bugzilla.redhat.com/show_bug.cgi?id=1629853